update
Working inventory page
