+++ Group9FinalProject/Group9FinalProject/Group9FinalProject.csproj
@@ -54,6 +54,8 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="clsInventory.cs" />
+    <Compile Include="clsPopulateInventoryPg.cs" />
     <Compile Include="clsSearch.cs" />
     <Compile Include="InventoryWindow.xaml.cs">
       <DependentUpon>InventoryWindow.xaml</DependentUpon>
