+++ Group9FinalProject/Group9FinalProject/InventoryWindow.xaml
@@ -1,12 +1,72 @@
-﻿<Window x:Class="Group9FinalProject.InventoryWindow"
+﻿<Window
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:Group9FinalProject"
+        xmlns:InvoiceDataSetTableAdapters="clr-namespace:Group9FinalProject.InvoiceDataSetTableAdapters" x:Class="Group9FinalProject.InventoryWindow"
         mc:Ignorable="d"
-        Title="InventoryWindow" Height="600" Width="800">
-    <Grid>
-        
+        Title="Update Inventory" Height="600" Width="800">
+    <Window.Resources>
+        <local:InvoiceDataSet x:Key="invoiceDataSet"/>
+        <CollectionViewSource x:Key="itemDescViewSource" Source="{Binding ItemDesc, Source={StaticResource invoiceDataSet}}"/>
+    </Window.Resources>
+    <Grid DataContext="{StaticResource itemDescViewSource}" Margin="0,0,3.6,0.4">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="151*"/>
+            <ColumnDefinition Width="639*"/>
+        </Grid.ColumnDefinitions>
+        <Button x:Name="btnDeleteItem" Content="Delete Item" HorizontalAlignment="Left" Margin="519.2,503,0,0" VerticalAlignment="Top" Width="96" Height="43" Background="#FFD9D0FF" Click="btnDeleteItem_Click" Grid.Column="1"/>
+        <Button x:Name="btnSave" Content="Save Changes" HorizontalAlignment="Left" Margin="254.2,503,0,0" VerticalAlignment="Top" Width="96" Height="43" Background="#FF92E696" Click="btnSave_Click" Grid.Column="1"/>
+        <Button x:Name="btnAddInventory" Content="Add Item" HorizontalAlignment="Left" Margin="16,503,0,0" VerticalAlignment="Top" Width="96" Height="43" Background="#FF92E696" Click="btnAddInvoice_Click"/>
+        <Label Content="Item Code:" Margin="16,41,59.8,495" />
+        <ComboBox x:Name="cboInventoryCode" Margin="99,41,614,495" IsSynchronizedWithCurrentItem="False" Grid.ColumnSpan="2">
+            <ComboBox.DataContext>
+                <InvoiceDataSetTableAdapters:ItemDescTableAdapter/>
+            </ComboBox.DataContext>
+        </ComboBox>
+        <DataGrid x:Name="dgInventory" AutoGenerateColumns="False" EnableRowVirtualization="True" ItemsSource="{Binding}" Margin="16,97,10,153" RowDetailsVisibilityMode="VisibleWhenSelected" Grid.ColumnSpan="2">
+            <DataGrid.Columns>
+                <DataGridTextColumn x:Name="itemCodeColumn" Binding="{Binding InventoryLetter}" Header="Item Code" Width="SizeToHeader"/>
+                <DataGridTextColumn x:Name="itemDescColumn" Binding="{Binding ItemDesc}" Header="Item Desc" Width="SizeToHeader"/>
+                <DataGridTextColumn x:Name="costColumn" Binding="{Binding ItemCost}" Header="Cost" Width="SizeToHeader"/>
+            </DataGrid.Columns>
+        </DataGrid>
+        <TextBlock x:Name="tbItemCodeLabel" HorizontalAlignment="Left" Margin="18,431,0,0" TextWrapping="Wrap" Text="Item Code:" VerticalAlignment="Top" Height="21" Width="78"/>
+        <TextBlock x:Name="tbItemDescLabel" HorizontalAlignment="Left" Margin="24.2,431,0,0" TextWrapping="Wrap" Text="Item Description:" VerticalAlignment="Top" Width="109" Grid.Column="1"/>
+        <TextBlock x:Name="tbItemCostLabel" HorizontalAlignment="Left" Margin="200.2,431,0,0" TextWrapping="Wrap" Text="Item Cost:" VerticalAlignment="Top" Width="109" Grid.Column="1"/>
+        <Grid x:Name="grid1" HorizontalAlignment="Left" Margin="83.2,46,0,0" VerticalAlignment="Top" Grid.Column="1">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto"/>
+                <ColumnDefinition Width="Auto"/>
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto"/>
+            </Grid.RowDefinitions>
+        </Grid>
+        <Grid x:Name="grid2" HorizontalAlignment="Left" Margin="46.2,459,0,0" VerticalAlignment="Top" Grid.Column="1">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto"/>
+                <ColumnDefinition Width="Auto"/>
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto"/>
+            </Grid.RowDefinitions>
+        </Grid>
+        <Grid x:Name="grid3" HorizontalAlignment="Left" Margin="350.2,458,0,0" VerticalAlignment="Top" Grid.Column="1">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto"/>
+                <ColumnDefinition Width="Auto"/>
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto"/>
+            </Grid.RowDefinitions>
+        </Grid>
+        <TextBox x:Name="tbItemCodeInp" HorizontalAlignment="Left" Height="23" Margin="18,452,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"/>
+        <TextBox x:Name="tbItemDescInp" HorizontalAlignment="Left" Height="23" Margin="24.2,454,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="142" Grid.Column="1"/>
+        <TextBox x:Name="tbItemCostInp" HorizontalAlignment="Left" Height="23" Margin="200.2,454,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="120" Grid.Column="1"/>
+        <Button x:Name="btnEditItem" Content="Edit Item" HorizontalAlignment="Left" Margin="147,503,0,0" VerticalAlignment="Top" Width="96" Height="43" Background="#FF92E696" Click="btnEditInvoice_Click" Grid.ColumnSpan="2"/>
+        <Button x:Name="btnClearText" Content="Clear" HorizontalAlignment="Left" Margin="125.2,503,0,0" VerticalAlignment="Top" Width="96" Height="43" Background="#FF92E696" Click="btnClear_Click" Grid.Column="1"/>
+
     </Grid>
 </Window>
